refactor(drawer): extract interaction test to dedicated file
